Lots of changes to mapmaker and game, added new enemytype and enemyspawner

+++ Tile.h
@@ -8,7 +8,7 @@ private:
 	//Core
 	bool isCollider;
 	bool isDamaging;
-	int damage;
+	double damage;
 
 	//Positioning
